README Documentation Typo Fix: 'space' -> 'comma'
Several examples in the README documentation mention escaping a space character, but then show escaping a comma character instead. This is a quick fix for those.

Note that I haven't gone through other files in the repo to check if they have that same or similar typos, so this may need adjusted elsewhere as well if this phrase was copy/pasted around in other documentation.